+++ nomenclate/core/nomenclature.py
@@ -8,6 +8,9 @@ import datetime
 import dateutil.parser as p
 import nomenclate.core.configurator as config
 import nomenclate.core.exceptions as exceptions
+from nomenclate.core.nomenclative import (
+    InputRenderer
+)
 from nomenclate.core.nlog import (
     getLogger,
     DEBUG
@@ -70,9 +73,7 @@ class TokenAttrDictHandler(object):
 
     @property
     def token_attrs(self):
-        for name, value in iteritems(self.__dict__):
-            if isinstance(value, TokenAttr):
-                yield value
+        return self.gen_object_token_attributes(self)
 
     @property
     def state(self):
@@ -97,6 +98,7 @@ class TokenAttrDictHandler(object):
     def build_name_attrs(self):
         """ Creates all necessary name attributes for the naming convention
         """
+        self.LOG.info('Building name attributes from default format order %s' % self.nom.format_order)
         self.purge_invalid_name_attrs()
         self.set_token_attrs({token: "" for token in self.nom.format_order})
 
@@ -119,8 +121,10 @@ class TokenAttrDictHandler(object):
             token_attr.set('')
 
     def set_token_attrs(self, input_dict):
+        self.LOG.info('Setting token attributes %s' % input_dict)
         for input_attr_name, input_attr_value in iteritems(input_dict):
             self.set_token_attr(input_attr_name, input_attr_value)
+        self.purge_invalid_name_attrs()
         self.update_nomenclate_token_attributes()
 
     def set_token_attr(self, token, value):
@@ -146,16 +150,21 @@ class TokenAttrDictHandler(object):
         self.__dict__[token.lower()] = TokenAttr(label=value, token=token)
 
     def update_nomenclate_token_attributes(self):
+        self.LOG.info('Updating nomenclate attributes - %s' % list(self.token_attrs))
         for token_attribute in self.token_attrs:
             setattr(self.nom, token_attribute.token, token_attribute)
+        self.clear_nomenclate_excess_token_attributes()
 
     def clear_nomenclate_excess_token_attributes(self):
-        for token_attr_key, token_attr_object in self.get_nomenclate_token_attributes():
+        for token_attr_key, token_attr_object in self.gen_object_token_attributes(self.nom):
             if token_attr_key not in self.nom.format_order:
                 delattr(self.nom, token_attr_key)
 
-    def get_nomenclate_token_attributes(self):
-        return [(attr, value) for attr, value in iteritems(self.nom.__dict__) if isinstance(value, TokenAttr)]
+    @staticmethod
+    def gen_object_token_attributes(object):
+        for name, value in iteritems(object.__dict__):
+            if isinstance(value, TokenAttr):
+                yield value
 
     @staticmethod
     def _validate_name_in_format_order(name, format_order):
@@ -180,329 +189,6 @@ class TokenAttrDictHandler(object):
         return ' '.join(['%s:%s' % (token_attr.token, token_attr.label) for token_attr in self.get_token_attrs()])
 
 
-class Nomenclative(object):
-    LOG = getLogger(__name__, level=DEBUG)
-
-    def __init__(self, input_str):
-        self.str = input_str
-        self.token_matches = []
-
-    def process_matches(self):
-        build_str = self.str
-        for token_match in self.token_matches:
-            if token_match.match == build_str[token_match.start:token_match.end]:
-                self.LOG.debug('%s - %s - %s - %s' % (token_match.match, list(token_match.sub),
-                                                      token_match.sub, build_str))
-                build_str = build_str[:token_match.start] + token_match.sub + build_str[token_match.end:]
-                self.adjust_other_matches(token_match)
-        return build_str
-
-    def adjust_other_matches(self, adjuster_match):
-        for token_match in [token_match for token_match in self.token_matches if token_match != adjuster_match]:
-            try:
-                token_match.adjust_position(adjuster_match)
-            except IndexError:
-                pass
-
-    def add_match(self, regex_match, substitution):
-        token_match = TokenMatch(regex_match, substitution)
-        try:
-            self.validate_match(token_match)
-            self.token_matches.append(token_match)
-        except IndexError:
-            self.LOG.warning('Not adding match %s as it conflicts with a preexisting match' % token_match)
-
-    def validate_match(self, token_match_candidate):
-        for token_match in self.token_matches:
-            try:
-                token_match.overlaps(token_match_candidate)
-            except exceptions.OverlapError as e:
-                self.LOG.error(e.message)
-
-    def __str__(self):
-        return '%s:%s' % (self.str, '\n'.join(self.token_matches))
-
-
-class TokenMatch(object):
-    def __init__(self, regex_match, substitution, group_name='token'):
-        self.match = regex_match.group(group_name)
-        self.start = regex_match.start(group_name)
-        self.end = regex_match.end(group_name)
-        self.sub = substitution
-        self.span = self.end - self.start
-
-    def adjust_position(self, other, adjust_by_sub_delta=True):
-        if isinstance(other, self.__class__):
-            if not other.overlaps(self) and self > other:
-                adjustment = other.span - len(other.sub) if adjust_by_sub_delta else other.span
-                self._adjust_order(adjustment)
-            else:
-                raise IndexError('Current TokenMatch overlaps with input TokenMatch\n\t%s\n\t%s' % (repr(self), repr(other)))
-        else:
-            raise IOError('Only TokenMatch objects are valid inputs to adjust_position')
-
-    def _adjust_order(self, adjust_value):
-        self.start -= adjust_value
-        self.end -= adjust_value
-
-    def overlaps(self, other):
-        value = self in other or other in self
-        if value:
-            return True
-        else:
-            raise exceptions.OverlapError('Match with overlaps with an existing match:\n\t%s\n\t%s' % (self, other))
-
-    def __contains__(self, other):
-        try:
-            return (self.start < other.start < self.end or self.start < other.end < self.end)
-        except:
-            raise NotImplementedError(
-                '{C} objects do not handle in syntax for non class objects'.format(C=self.__class__.__name__))
-
-    def __eq__(self, other):
-        try:
-            return (self.start == other.start and self.end == other.end and
-                    self.match == other.match and self.sub == other.sub)
-        except:
-            raise NotImplementedError(
-                '{C} objects do not handle == syntax for non class objects'.format(C=self.__class__.__name__))
-
-    def __lt__(self, other):
-        if isinstance(other, self.__class__):
-            return self.end <= other.start
-        else:
-            raise NotImplementedError
-
-    def __gt__(self, other):
-        if isinstance(other, self.__class__):
-            return self.start >= other.end
-        else:
-            raise NotImplementedError
-
-    def __str__(self):
-        return '%s (%d)- [%d:%d] - replacement = %s' % (self.match, self.span, self.start, self.end, self.sub)
-
-
-class InputRenderer(type):
-    RENDER_FUNCTIONS = {}
-    REGEX_PARENTHESIS = r'([\(\)]+)'
-    REGEX_BRACKETS = r'([\{\}]+)'
-    REGEX_STATIC_TOKEN = r'(\(\w+\))'
-    REGEX_BRACKET_TOKEN = r'(\{\w+\})'
-    LOG = getLogger(__name__, level=DEBUG)
-
-    def __new__(mcs, name, bases, dct):
-        cls = type.__new__(mcs, name, bases, dct)
-        token = dct.get('token', None)
-        if token:
-            mcs.RENDER_FUNCTIONS[token] = cls
-        return cls
-
-    @classmethod
-    def render_unique_tokens(cls, nomenclate_object, input_dict):
-        for k, v in iteritems(input_dict):
-            renderer = None
-            for token, render_method in iteritems(cls.RENDER_FUNCTIONS):
-                if token in k:
-                    renderer = render_method
-            if 'render' in dir(renderer):
-                cls.LOG.info('render_unique_tokens() - token %s renderer %s with token settings %s' % (k, v, nomenclate_object.get_token_settings(token)))
-                input_dict[k] = renderer.render(v, nomenclate_object, **nomenclate_object.get_token_settings(token))
-
-    @classmethod
-    def render_nomenclative(cls, nomenclate_object):
-        nomenclative = Nomenclative(nomenclate_object.format_string)
-        token_values = nomenclate_object.token_dict.token_attr_dict
-        cls.render_unique_tokens(nomenclate_object, token_values)
-        rendered_nomenclative = nomenclate_object.format_string_object.format_string
-
-        cls._prepend_token_match_objects(token_values, rendered_nomenclative)
-
-        for token, match_value in iteritems(token_values):
-            nomenclative.add_match(*match_value)
-
-        rendered_nomenclative = cls.cleanup_formatted_string(nomenclative.process_matches())
-        return rendered_nomenclative
-
-    @classmethod
-    def _prepend_token_match_objects(cls, token_values, incomplete_nomenclative):
-        for token, value in iteritems(token_values):
-            re_token = r'(?P<token>((?<![a-z]){TOKEN})|((?<=[a-z]){TOKEN_CAPITALIZED}))'
-            re_token = re_token.format(TOKEN=token,
-                                       TOKEN_CAPITALIZED=token[0].upper()+token[1:])
-            re_matches = re.finditer(re_token, incomplete_nomenclative, 0)
-
-            for re_match in re_matches:
-                token_values[token] = (re_match, value)
-
-        cls._clear_non_matches(token_values)
-
-    @staticmethod
-    def _clear_non_matches(token_values):
-        to_delete = []
-        for token, value in iteritems(token_values):
-            if isinstance(value, str) or not isinstance(value, tuple):
-                to_delete.append(token)
-
-        for delete in to_delete:
-            token_values.pop(delete)
-
-    @staticmethod
-    def _render_replacements(replacements_dict, incomplete_nomenclative):
-        for token, match_value in iteritems(replacements_dict):
-            match, value = match_value
-            incomplete_nomenclative = incomplete_nomenclative.replace(match, value)
-
-    @classmethod
-    def cleanup_formatted_string(cls, formatted_string):
-        """ Removes unused tokens/removes surrounding and double underscores
-        Args:
-            formatted_string (string): string that has had tokens replaced
-        Returns (string): cleaned up name of object
-        """
-        # TODO: chunk this out to sub-processes for easier error checking, could be own class
-        # Remove whitespace
-        result = formatted_string.replace(' ', '')
-        # Remove any static token parentheses
-        result = re.sub(cls.REGEX_PARENTHESIS, '', result)
-        # Remove any multiple underscores
-        result = re.sub('_+', '_', result)
-        # Remove trailing or preceding underscores
-        result = re.match(r'^_*(.*?)_*$', result)
-        if result:
-            return result.groups()[0]
-        else:
-            return result
-
-    @staticmethod
-    def _get_alphanumeric_index(query_string):
-        """ Given an input string of either int or char, returns what index in the alphabet and case it is
-        Args:
-            query_string (str): query string
-        Returns [int, str]: list of the index and type
-        """
-        #TODO: could probably rework this. it works, but it's ugly as hell.
-        try:
-            return [int(query_string), 'int']
-        except ValueError:
-            if len(query_string) == 1:
-                if query_string.isupper():
-                    return [string.uppercase.index(query_string), 'char_hi']
-                elif query_string.islower():
-                    return [string.lowercase.index(query_string), 'char_lo']
-            else:
-                raise IOError('The input is a string longer than one character')
-        return [0, 'char_hi']
-
-
-class RenderBase(object):
-    __metaclass__ = InputRenderer
-    LOG = getLogger(__name__, level=DEBUG)
-
-    token = None
-
-    @classmethod
-    def render(cls, arg, nomenclate_object, **kwargs):
-        raise NotImplementedError
-
-
-class RenderDate(RenderBase):
-    token = 'date'
-
-    @classmethod
-    def render(cls, date, nomenclate_object, **kwargs):
-        if date == 'now':
-            d = datetime.datetime.now()
-        else:
-            try:
-                d = p.parse(date)
-            except ValueError:
-                return ''
-
-        return d.strftime(get_keys_containing('format', nomenclate_object.__dict__, '%Y-%m-%d'))
-
-
-class RenderVar(RenderBase):
-    __metaclass__ = InputRenderer
-    token = 'var'
-
-    @classmethod
-    def render(cls, var, nomenclate_object, **kwargs):
-        if var:
-            var_index = kwargs.get('%s_index' % cls.token, 0)
-            return cls._get_variation_id(var_index, var.isupper())
-        else:
-            return var
-
-    @staticmethod
-    def _get_variation_id(integer, capital=False):
-        """ Convert an integer value to a character. a-z then double aa-zz etc
-        Args:
-            integer (int): integer index we're looking up
-            capital (bool): whether we convert to capitals or not
-        Returns (str): alphanumeric representation of the index
-        """
-        # calculate number of characters required
-        base_power = base_start = base_end = 0
-        while integer >= base_end:
-            base_power += 1
-            base_start = base_end
-            base_end += pow(26, base_power)
-        base_index = integer - base_start
-
-        # create alpha representation
-        alphas = ['a'] * base_power
-        for index in range(base_power - 1, -1, -1):
-            alphas[index] = chr(97 + (base_index % 26))
-            base_index /= 26
-
-        characters = ''.join(alphas)
-        return characters.upper() if capital else characters
-
-
-class RenderVersion(RenderBase):
-    __metaclass__ = InputRenderer
-    token = 'version'
-
-    @classmethod
-    def render(cls, version, nomenclate_object, **kwargs):
-        padding = kwargs.get('%s_padding' % cls.token, 4)
-        return '%0{0}d'.format(padding) % version
-
-
-class RenderType(RenderBase):
-    __metaclass__ = InputRenderer
-    token = 'type'
-
-    @classmethod
-    def render(cls, type, nomenclate_object, **kwargs):
-        length = kwargs.get('%s_length' % cls.token, None)
-        try:
-            suffixes = nomenclate_object.cfg.get(nomenclate_object.SUFFIXES_PATH, return_type=dict)
-        except exceptions.ResourceNotFoundError:
-            cls.LOG.warning('Could not find entry, defaulting to current: %s' % type)
-        options = gen_dict_key_matches(type, suffixes)
-        cls.LOG.debug('type: %s = %s' % (length, list(options)))
-        return options
-
-
-class RenderSide(RenderBase):
-    __metaclass__ = InputRenderer
-    token = 'side'
-
-    @classmethod
-    def render(cls, side, nomenclate_object, **kwargs):
-        length = kwargs.get('%s_length' % cls.token, None)
-        side_options = nomenclate_object.cfg.get(nomenclate_object.OPTIONS_PATH + [cls.token, side])
-        cls.LOG.debug('side: %s' % list(side_options))
-
-        for side_option in side_options:
-            if len(side_option) == length:
-                return side_option
-
-        return next(iter(side_options))
-
-
 class FormatString(object):
     FORMAT_STRING_REGEX = r'([A-Za-z][^A-Z_.]*)'
     FORMAT_STRING_REGEX = r'(?:(?<=\()[\w]+(?=\)))|([A-Za-z0-9][^A-Z_.\(\)]+)'
@@ -520,6 +206,8 @@ class FormatString(object):
             self._validate_format_string(format_target)
             self.format_string = format_target
             self.format_order = self.get_format_order(format_target)
+            self.LOG.info('Successfully set format string: %s and format order: %s' % (self.format_string,
+                                                                                       self.format_order))
         except exceptions.FormatError as e:
             msg = "Could not validate input format target %s"
             self.LOG.error('%s, %s' % (msg, e.message))
@@ -628,15 +316,16 @@ class Nomenclate(object):
                                               e.g. - this_is_a_naming_string
         Returns None: raises IOError if failure
         """
+        self.LOG.info('initialize_format_options with %s' % format_target)
         try:
             format_path = format_target if isinstance(format_target, list) else [format_target]
-            format_target = self.cfg.search(format_path, full_path=True)
-            #format_target = self.cfg.get(format_path, return_type=str)
+            format_target = self.cfg.get(format_path, return_type=str)
 
         except (exceptions.ResourceNotFoundError, StopIteration):
             format_target = format_target or self.cfg.get(self.DEFAULT_FORMAT_PATH, return_type=str)
 
         finally:
+            self.LOG.info('format target is now %s after looking in the config for a match' % format_target)
             self.swap_format(format_target)
 
     def initialize_options(self):
@@ -649,6 +338,8 @@ class Nomenclate(object):
         pass
 
     def swap_format(self, format_target):
+        self.LOG.info('Attempting to swap format to target %s' % format_target)
+        format_target = self.cfg.get(format_target, return_type=str)
         self.format_string_object.swap_format(format_target)
 
     def get(self, **kwargs):
@@ -713,14 +404,17 @@ class Nomenclate(object):
             Used to weed out config keys from tokens in a given input.
         """
         configs = {}
-
+        to_pop = []
         for k, v in iteritems(input_dict):
             try:
                 self.cfg.get(self.CONFIG_PATH + [k])
-                input_dict.pop(k, None)
+                to_pop.append(k)
                 configs[k] = v
             except exceptions.ResourceNotFoundError:
                 pass
+
+        for pop in to_pop:
+            input_dict.pop(pop, None)
 
         self.initialize_config_settings(input_dict=configs)
